pam_slurm_save_cgroups: delete_only_on_close, as_type

+++ pam_slurm_save_cgroups.sh
@@ -4,7 +4,7 @@
 #
 #   pam_slurm_save_cgroups.sh
 #
-#   Copyright (C) 2018-2020 Hebrew University of Jerusalem Israel, see LICENSE
+#   Copyright (C) 2018-2022 Hebrew University of Jerusalem Israel, see LICENSE
 #   file.
 #
 #   Author: Yair Yarom <irush@cs.huji.ac.il>
@@ -12,6 +12,7 @@
 ################################################################################
 
 set -e
+set -u
 
 _savedir="/run/slurm-saved-cgroups/"
 if [[ ! -e "$_savedir" ]]; then
@@ -25,14 +26,37 @@ if [[ ( ! -d "$_savedir" ) || ( `stat --format '%F/%a/%u/%g' "$_savedir"` != 'di
     exit 3
 fi
 
+action=
+as_type=${PAM_TYPE}
+delete_only_on_close=
+while [[ "$#" -gt 0 ]]; do
+    arg=$1
+    shift;
+    case $arg in
+        as_type=*)
+            as_type=`echo $arg | cut -d= -f2-`
+            ;;
+        save|restore|slurm)
+            action=$arg
+            ;;
+        delete_only_on_close)
+            delete_only_on_close=1
+            ;;
+        *)
+            echo "Bad argument: $arg" 1>&2
+            exit 3;
+    esac
+done
+
+
 _ppid=`awk '$1=="PPid:"{print $2}' /proc/$$/status`
 if [[ -z "$_ppid" ]]; then
    echo "Don't know my PPID" 1>&2
    exit 1
 fi
-_savefile="$_savedir/${PAM_SERVICE}.${PAM_TYPE}.${PAM_USER}.${_ppid}"
+_savefile="$_savedir/${PAM_SERVICE}.${as_type}.${PAM_USER}.${SLURM_JOB_ID:-$_ppid}"
 
-case "$1" in
+case "$action" in
     save)
         cat /proc/$_ppid/cgroup > $_savefile
     ;;
@@ -66,6 +90,10 @@ case "$1" in
     ;;
 
     slurm)
+        if [[ ! -e ${_savefile} ]]; then
+            exit 200
+        fi
+
         # first lets restore
         for cgroup in `awk -F: '$2&&$3~/^\/slurm\//&&$2!~/name=systemd/ {printf "/sys/fs/cgroup/%s%s/tasks\n", $2, $3}' ${_savefile}`; do
             echo $_ppid >> $cgroup
@@ -115,10 +143,12 @@ case "$1" in
 
         done
 
-        rm "${_savefile}"
+        if [[ "${PAM_TYPE}" = close_session || "x${delete_only_on_close}" = "x" ]]; then
+            rm "${_savefile}"
+        fi
     ;;
     *)
-        echo "Either save or restore" 1>&2
+        echo "Unknown action \"${action}\"" 1>&2
         exit 2
 esac
